requirements.txt
Django==1.11.4
olefile==0.44
Pillow==4.3.0
pytz==2017.2
